Move all pages outside of the routing App file. Change signup and login so that it is two separate forms.
